+++ src/orc/__init__.py
@@ -1,5 +1,5 @@
 """Components of models."""
 
-from orc import drivers, embeddings, models, rc, readouts, utils
+from orc import drivers, embeddings, models, rc, readouts, utils, data
 
-__all__ = ["drivers", "embeddings", "models", "readouts", "utils", "rc"]
+__all__ = ["drivers", "embeddings", "models", "readouts", "utils", "rc", "data"]
