+++ src/orc/data/__init__.py
@@ -1,0 +1,6 @@
+"""Implementations of data generation and handling functions."""
+
+from orc.data import integrators
+from orc.data.integrators import *
+
+__all__ = ["integrators"]
